delete plant moved to edit plant page
