+++ src/components/Loading.js
@@ -3,7 +3,7 @@ import React from 'react'
 export const Loading = () => {
   return (
     <div>
-       LOADING......
+      <p>LOADING......</p>
     </div>
   )
 }
